Fix possible startup race conditions when multiple workers are present.
If multiple workers were in use, it was possible the "storage" item
never got set, since it was wrapped in a lock that exited immediately if
any other workers were still in progress performing the initial setup.

So shift that initial setup logic to the master process as an easy way
of ensuring it only happens once.

Add explicit tests for multiple workers.