Better error handling for background tasks.
Ensure that if pcall fails on a function call, we still unlock the
lock (previously it was possible that if a renewal failed, the lock
would remain in place indefinitely).

Also assign long expiration times to each lock to ensure that even in
the event of a hard worker process crash, things don't stay locked
indefinitely.