Fix recovering certs from Let's Encrypt's files if deleted from storage.
In the edge-case that the certs are deleted from the storage (or never
made it into the storage), then we attempt to restore the certificates
from the Let's Encrypt files if those exist. However, this logic broke
when we upgraded to letsencrypt.sh v0.2.0
(https://github.com/GUI/lua-resty-auto-ssl/pull/7), due to the hook
script having a different expected function signature.

This also adds a test case for this scenario to make sure it continues
to work as expected.