Fix handling of long command line strings being sent to resty-shell.
This cropped up in our CI environment, since due to the recent changes
to the letsencrypt_hooks arguments, we needed to pass additional command
line arguments in. This worked fine in my local environment, but this
ended up breaking on our CI environment. This uncovered the fact that
sockproc limits the length of the command line string. Since some of the
file paths happened to be longer in the CI environment, we ended up
exceeding this limit.

This resolves it by using a fork of sockproc to allow for longer command
line strings: https://github.com/juce/sockproc/pull/5