Refactor the verification of sockproc status
I have added here an OR logical operator to verify sockproc state (exit_code) against a boolean value (in this case true).

This became necessary from the context when I noticed that exit_code return a boolean true which did not pass the conditional test on line 6 of start_sockproc.lua.

See this Issue for more details #16