Fix file permissions under LuaRocks 2.4. Add Docker test setup.
The executable files installed (like sockproc) were becoming
un-executable if installed with LuaRocks 2.4. LuaRocks 2.4 now sets
explicit permissions on the installed files, so that all executables
must be treated as "bin" files. This means we need to reorganize our
file structure a bit. See: https://github.com/luarocks/luarocks/pull/549
and https://github.com/GUI/lua-resty-auto-ssl/issues/47

This also adds a Docker setup for running tests, to hopefully make the
test environment a bit easier to get up and running.

As part of the new test setup, this also upgrades the test
dependencies (which is how I stumbled into the LuaRocks 2.4 issue),
fixes multiple_workers.t so it works when run as root, and improves some
error handling in the tests.