Log errors if shdict runs out of memory and items get removed.
This should more generally help with potential issues seen in
https://github.com/GUI/lua-resty-auto-ssl/issues/48 where old items
might get removed from the shdict if there's too many certificates for
the configured amount of memory. By logging this as an error, it will
hopefully make debugging and alerting to the real issue easier (need to
increase shdict memory).

Since we don't store duplicate certs for domains, and the certificate
data is stored with a TTL, I think this should be okay to log at the
error level, since if this happens, thing's aren't going well, since
there's not enough memory for active domains.