Fix sockproc possibly inheriting nginx's sockets/file descriptors.
When we launch the sockproc process from inside nginx, we explicitly try
to clear all extra file descriptors, so nginx doesn't inherit the nginx
processes sockets. If sockproc does inherit these sockets, then if nginx
gets stopped or restarted, sockproc takes over listening on those
sockets, which means nginx can't be started again.

However, the existing implementation of clearing these file descriptors
could fail under certain circumstances, leading to nginx being unable to
restart. It depends on how nginx is started, but this could reliably
fail under this scenario:

- nginx was started under an init script.
- nginx had worker processes running as a less privileged user.
- The user that originally ran the init start command was "cd"ed into a
  directory that the less privileged nginx worker user didn't have
  permissions to be in.

This was failing due to the `find` command being combined with `-exec
basename` to find the file descriptors to clear. `basename` would fail
if the current user didn't have permissions to be in the current
directory (even if it was only finding the basename of paths in /proc
which it did have permissions to). This led to no file descriptors being
cleared. The `find` command would also generate a warning message after
finishing about not being able to restore the initial directory.

To fix this:

- Our script to start sockproc now explicitly sets the current directory
  to `/`, so any commands that require permissions to be in the current
  directory should work.
- Extract the basename of the /proc/ paths using bash, instead of
  shelling out to the `basename` command.
- If for any reason, clearing the file descriptors based on /proc/ paths
  fails, also fallback to the more naive OS x approach in Linux too.

A test has been added to the test suite to test this behavior. Since the
Makefile was getting a little unwieldy (particularly with another test
that needed all the "worker perms" environment variables), shift the
commands used to run the tests into a wrapper shell script to make
things a bit easier to manage.