Fix intermittent test issues (or at least make debugging easier).
We're still seeing some rare intermittent test failures if running
repeatedly. But some of the failures don't make sense based on the CI
logs (it's reporting failures for error log output that isn't in the
log file). But I think the issue might be related to repeated runs, and
not seeing the log files for earlier runs, so disable repeated runs to
see if this fixes things, or at least makes debugging easier when things
do fail.