Fix file descriptor tests for CI environment.
I'm not exactly sure why, but sometimes in the CI environment, the
STDERR descriptor is listed as the nginx log file, instead of /dev/null.
I think this should be fine, we just need to normalize our tests to
handle both situations (but for some reason this doesn't seem to ever
happen in the Docker test environment, but does happen sporadically in
the CI environment).