Increase the suggested body buffer size for the HTTP hook server.
This eliminates potential issues if the HTTP POST size to the hook
server containing the cert exceeded nginx's default buffer size (which
would prevent the hook server from being able to parse the POST args):
https://github.com/GUI/lua-resty-auto-ssl/issues/65

Based on some quick tests, it looks like the POST to `/deploy-cert`,
containing the certificate chain and private key was the largest POST.
These look to be in the neighborhood of 10KB, while nginx's default
`client_body_buffer_size` might be either 8KB or 16KB depending on the
exact system architecture. To address this, increase the suggested
configuration in the README to 128KB (which is probably overkill, but
provides plenty of space in case Let's Encrypt's full certificate chain
ever becomes bigger).

This also adds some better error logging and error handling to the hook
server, and adds more specific tests around the hook server.