Fix error handling for `get_cert` and not found handling for Redis.
The storage's `get_cert` method was not returning the right number of
arguments when an error was returned by a storage adapter's `get`
method. This could lead to inconsistent error handling, since the
returned error was not in the expected place.

Fixing this uncovered a difference between between how the redis and
file storage adapters were handling not found results. The file adapter
was just returning nil with no error, but the redis adapter was
returning "not found" as an error. Fix this so that the redis adapter
behaves like the file adapter and no longer returns "not found" results
as an error.