Refactor "get_cert" and "issue_cert" functions to return tables of data.
This refactors some of the internal functions to return data in a single
table instead of as multiple return values.

With the recent addition of return both expiry details and error
handling to the "get_certs" method, these ended up conflicting since it
changed the position of the results. Since returning this many values
from a single function was getting a bit unwieldy, it seemed best to
refactor things so that all the cert methods return and pass along the
cert data as a single table, rather than as positional
returns/arguments. This will also make it easier if we add more metadata
to the cert data in the future, since the return signature can now
remain the same (table, err), rather than needing to worry about the
exact position of return values.