Shift ordering of allow_domain lookup.
This tweaks the changes made in https://github.com/GUI/lua-resty-auto-ssl/pull/107

This shifts the allow_domain callback to happen after the cert lookup in
memory, but before the storage lookup. Since storage lookups could also
be costly, this tries to strike a middle-ground approach to when
allow_domain is performed.