Update legacy certificates in storage with expiration date
af15771 and 3b0038d added the field "expiry" to the certificate, which speeds up the renewal check considerably for certificates that have this information set. For existing ones that don't, use the most recent backup timestamp to establish the expiration date, and add this information to storage, so that future renewal jobs are faster for those as well.

This is in preparation for the following patch in the series, which uses the "expiry" field to purge certificates where renewal has failed repeatedly.